chroot_version_hooks.d: drop support for upgrading SDKs pre-2019
Punt a lot of old upgrade hooks that only get used for SDKs that were
created before 2019 and have not been upgraded since.  The tooling will
already detect this scenario (because the upgrade hooks are missing) and
throw an error telling the user to recreate the SDK.

We know that people aren't doing this because apparently it's been broken
for sometime now when trying to upgrade with an old version of portage.
Since no one has complained, rather than try and fix a scenario that no
one is testing, just drop it all.

BUG=None
TEST=precq passes

Change-Id: I2aeea1a0535d94bf937e4fe3464a046a6e87e0be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/2028989
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>