scripts: Build DLC images during the build_image phase.
Previously, the DLC images were created at the build_packages phase.
With this change, the build_packages will only build the DLC packages,
and the DLC images will be created at the build_images phase.
The change in common.sh is to exclude the DLC files from the rootfs
during emerge to the image and DUT, but ideally, the DLC files should
not be included in tha tarball during the build_package phase at all.

BUG=chromium:1050646
TEST=build_packages, build test image, cros flash local image.

Cq-Depend: chromium:2067295
Change-Id: Ifb827e101d9a1d65a1d1f5b2c6c58b226d240a7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/2067939
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>