enter_chroot: move C.UTF-8 mangling before chroot usage
We convert C.UTF-8 to en_US.UTF-8, but we did it before we tried to
enter the SDK & initialize the env.  Move that earlier so we pass in
the right setting.

This also means we can delete the explicit check for C.UTF-8 which
is breaking on builders using that locale.

BUG=chromium:1052082, chromium:1069678
TEST=`LC_ALL=C.UTF-8 cros_sdk` works

Change-Id: I68473b9c2bab3522090249db9e10adbac9746a14
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/2145030
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>