install hooks: disable compiler flags checks for now
We have a check in qa-elf.sh that looks at the .GCC.command.line section
to detect what compiler flags were used. However, we do not use the
-frecord-gcc-switches option that enables generating this section,
so the check always fails. We have plans to update it, but, for now,
let's just disable it so we don't spam developers with useless error
messages.

BUG=chromium:1061666
TEST=emerge-${BOARD} app-arch/tar; \
  check that "could not find section '.GCC.command.line'" errors are gone.

Change-Id: I3fe60c2f5ef8df87b8ad4b61e1f1e2a7707b8e99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/2208544
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>