enter_chroot: force sane PATH with locale-gen
If the host distro has a default PATH setup that doesn't match what
we use inside the SDK, locale-gen might not be found, or programs it
expects to be in $PATH won't be.  Force the setting to match the SDK
so the host distro doesn't matter.

BUG=None
TEST=running cros_sdk with a reduced PATH used to fail but now works

Change-Id: I136bf9f7244d9c644c90ce939516a33b6c16436f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/2208936
Reviewed-by: Alexandru M Stan <amstan@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>