dev-install: do not require binpkgs for lists
When generating the list of available packages, we don't care about
the specific versions, just the $CATEGORY/$PN, so don't force the use
of binpkgs when building the list.  This way we can generate it even
when only building a base image without test packages.

BUG=chromium:1072761
TEST=`./build_image ... base` works

Change-Id: Iae81b92dc4b1bb0c30278dfc195bb360fb56b300
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/2213619
Reviewed-by: Michael Mortensen <mmortensen@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>