build_packages: Always rebuild consumers of active cros_workon packages
The first time a user installs a workon package the rebuild logic would
work correctly as the new 9999 version is enough information for Portage
to correctly rebuild the packages that consume the workon package, but
subsequent rebuilds would not be handled correctly as they would already
see the 9999 version as being installed.

Change this logic to always recompile anything that directly depends on
an active cros_workon package in order to ensure that a user's local
changes to the package are always correctly reflected in the contents of
the overall build.

BUG=chromium:1071530, chromium:1017842
TEST=`build_packages`
TEST=CQ

Change-Id: I37b6752541f9481667983b12d7f79e143dd3b1c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/2250552
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>