install hooks: Allow lib64 for google-breakpad
We want to build libbreakpad_client.a as a 64-bit archive so we can link
it with core_collect64 in crash-reporter. This sanity check fails though
because the package installs the archive to /usr/lib64. Skip this sanity
check when the package is google-breakpad so we can have both the 32-bit
and 64-bit versions of libbreakpad_client.a useable by crash-reporter.

BUG=b:164481201
TEST=USE="cros_arm64" emerge-trogdor google-breakpad

Change-Id: Ifaa60d3156cd106cde0b9a512bf5515d85e7d757
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/2358913
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>