build_image: rename dev-only-extras.tbz2
The tbz2 suffix makes portage unhappy and spams a lot of warnings
that in turn makes users unhappy.  Rename to a file portage won't
care about and use better compression in the process.

It's safe to rename as the build will skip a missing file atm.

BUG=chromium:448178
TEST=CQ passes

Change-Id: I49e6b48fabd3762703f4eecb2fd0d5df14407063
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/2365919
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Michael Mortensen <mmortensen@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>