create_remote_test_driver: add a script to make a buildable RTD image
This is roughly what we'll need our new Build API endpoint to invoke.
It'll poke into the SDK, generate a folder containing a Dockerfile and
supporting source files, then return that folder path back to the
builder. The builder will then somehow build the docker image.

I talked about this with Alex, and we agreed that it makes sense to just
have this as a shell script for now. We can move the whole thing into
chromite as a Python script if that becomes necessary. Even now, the
Build API can just call out to this script if it wants.

BUG=chromium:1149702
TEST=ran it locally

Exempt-From-Owner-Approval: got verbal approval from saklein for this approach
Change-Id: Idf59c0fcb330473006405eb423834ba382e40426
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/2551703
Commit-Queue: Sean Abraham <seanabraham@chromium.org>
Tested-by: Sean Abraham <seanabraham@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Jared Loucks <jaredloucks@google.com>
Reviewed-by: Brigit Rossbach <brigitr@google.com>