dev-install: restrict virtual/target-os to binpkg-only
We use the virtual/target-os list to determine what packages are baked
into the rootfs so we can remove them from uploads later on.  If we use
the full list when building from source, we will incorrectly think that
they're available in the rootfs, and thus not upload them for installing
later on.

BUG=chromium:1072761, chromium:1108379
TEST=built list of packages before & after and compared them

Change-Id: I41e31da98c590d0d4bca880bc541f2526716f331
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/2631694
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>