enter_chroot: remove /run mount
This was added to support /dev/shm on Ubuntu systems that create a
symlink from /dev/shm to /run/shm.  There is nothing else from /run
that we need, and a whole lot that we don't want to expose.  So drop
the /run mount entirely, and just create a /run/shm dir in case the
host distro needs it.

BUG=None
TEST=CQ passes

Change-Id: Ia0a89aa9ee953db8488e5629fd7a316d5ce9e93c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosutils/+/2940042
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>